Splitting up memory interface, continued project refactor
