+++ Squalr.Engine.Memory/Extensions.cs
@@ -12,9 +12,9 @@ namespace Squalr.Engine.Memory
         /// <param name="address">The address to read from.</param>
         /// <param name="success">Whether or not the read succeeded.</param>
         /// <returns>The value read from memory.</returns>
-        public static Object Read(this IVirtualMemoryAdapter memory, DataType elementType, IntPtr address, out Boolean success)
+        public static Object Read(this IMemoryReader memory, DataType elementType, IntPtr address, out Boolean success)
         {
-            return VirtualMemoryAdapterFactory.GetVirtualMemoryAdapter().Read(elementType, address, out success);
+            return Reader.Default.Read(elementType, address, out success);
         }
 
         /// <summary>
@@ -24,9 +24,9 @@ namespace Squalr.Engine.Memory
         /// <param name="address">The address to read from.</param>
         /// <param name="success">Whether or not the read succeeded.</param>
         /// <returns>The value read from memory.</returns>
-        public static T Read<T>(this IVirtualMemoryAdapter memory, IntPtr address, out Boolean success)
+        public static T Read<T>(this IMemoryReader memory, IntPtr address, out Boolean success)
         {
-            return VirtualMemoryAdapterFactory.GetVirtualMemoryAdapter().Read<T>(address, out success);
+            return Reader.Default.Read<T>(address, out success);
         }
     }
     //// End class
