+++ Squalr/Source/Snapshots/SnapshotManagerViewModel.cs
@@ -1,7 +1,6 @@
 ﻿namespace Squalr.Source.Snapshots
 {
     using GalaSoft.MvvmLight.CommandWpf;
-    using Squalr.Engine;
     using Squalr.Engine.Memory;
     using Squalr.Engine.Output;
     using Squalr.Properties;
@@ -234,10 +233,10 @@
             MemoryTypeEnum allowedTypeFlags = MemoryTypeEnum.None | MemoryTypeEnum.Private | MemoryTypeEnum.Image;
 
             IntPtr startAddress = IntPtr.Zero;
-            IntPtr endAddress = Eng.GetInstance().VirtualMemory.GetMaxUsermodeAddress().ToIntPtr();
+            IntPtr endAddress = Query.Default.GetMaxUsermodeAddress().ToIntPtr();
 
             List<ReadGroup> memoryRegions = new List<ReadGroup>();
-            IEnumerable<NormalizedRegion> virtualPages = Eng.GetInstance().VirtualMemory.GetVirtualPages(
+            IEnumerable<NormalizedRegion> virtualPages = Query.Default.GetVirtualPages(
                 requiredPageFlags,
                 excludedPageFlags,
                 allowedTypeFlags,
@@ -267,7 +266,7 @@
             if (SettingsViewModel.GetInstance().IsUserMode)
             {
                 startAddress = IntPtr.Zero;
-                endAddress = Eng.GetInstance().VirtualMemory.GetMaxUsermodeAddress().ToIntPtr();
+                endAddress = Query.Default.GetMaxUsermodeAddress().ToIntPtr();
             }
             else
             {
@@ -276,7 +275,7 @@
             }
 
             List<ReadGroup> memoryRegions = new List<ReadGroup>();
-            IEnumerable<NormalizedRegion> virtualPages = Eng.GetInstance().VirtualMemory.GetVirtualPages(
+            IEnumerable<NormalizedRegion> virtualPages = Query.Default.GetVirtualPages(
                 requiredPageFlags,
                 excludedPageFlags,
                 allowedTypeFlags,
@@ -298,7 +297,7 @@
         /// <returns>The created snapshot.</returns>
         private Snapshot CreateSnapshotFromModules()
         {
-            IEnumerable<ReadGroup> moduleGroups = Eng.GetInstance().VirtualMemory.GetModules().Select(region => new ReadGroup(region.BaseAddress, region.RegionSize));
+            IEnumerable<ReadGroup> moduleGroups = Query.Default.GetModules().Select(region => new ReadGroup(region.BaseAddress, region.RegionSize));
             Snapshot moduleSnapshot = new Snapshot(null, moduleGroups);
 
             return moduleSnapshot;
@@ -314,7 +313,7 @@
             Snapshot snapshot = this.CreateSnapshotFromUsermodeMemory();
 
             // Remove module regions
-            IEnumerable<ReadGroup> moduleGroups = Eng.GetInstance().VirtualMemory.GetModules().Select(region => new ReadGroup(region.BaseAddress, region.RegionSize));
+            IEnumerable<ReadGroup> moduleGroups = Query.Default.GetModules().Select(region => new ReadGroup(region.BaseAddress, region.RegionSize));
             snapshot.ReadGroups = snapshot.ReadGroups.Where(group => moduleGroups.All(moduleGroup => moduleGroup.BaseAddress != group.BaseAddress));
 
             return snapshot;
