+++ Squalr/View/PointerScanResults.xaml
@@ -7,7 +7,7 @@
     xmlns:d="http://schemas.microsoft.com/expression/blend/2008"
     xmlns:intr="clr-namespace:System.Windows.Interactivity;assembly=System.Windows.Interactivity"
     xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006"
-    xmlns:types="clr-namespace:Squalr.Engine.DataTypes;assembly=Squalr.Engine.DataTypes"
+    xmlns:types="clr-namespace:Squalr.Engine.DataTypes;assembly=Squalr.Engine"
     x:Name="SqualrPointerScanResults"
     d:DesignHeight="256"
     d:DesignWidth="512"
